pyproject.toml
[project]
name = "langchain-chat-with-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langchain-redis>=0.2.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
